pyproject.toml
[project]
name = "discrete-dino"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=3.5.0",
    "faiss-cpu>=1.10.0",
    "loguru>=0.7.3",
    "numpy>=2.2.5",
    "tqdm>=4.67.1",
]
